CreateProject Update
Adding a validation to check if the auth is valid, I was going to use a policy but I don't really want to overcomplicate it when I don't really think I'll be using this elsewhere.

End date must now also be after the start date or equal to it by using after or equal.